requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-groq==0.1.4
streamlit==1.35.0
python-dotenv==1.0.0
